Move C++ `FuncType` to its own file
