+++ crates/c-api/include/wasmtime/types.hh
@@ -1,6 +1,7 @@
 #ifndef WASMTIME_TYPES_HH
 #define WASMTIME_TYPES_HH
 
+#include <wasmtime/types/func.hh>
 #include <wasmtime/types/global.hh>
 #include <wasmtime/types/memory.hh>
 #include <wasmtime/types/table.hh>
